+++ assets/blocks/progress-bar-block/block.json
