+++ assets/blocks/progress-bar-block/index.js
