+++ assets/blocks/progress-bar-block/progress-bar-edit.js
@@ -36,7 +36,7 @@ import ProgressBar, {
  * @param {number}   props.attributes.borderRadius The border radius of the progress bar.
  * @param {Function} props.setAttributes           Callback to set the component attributes.
  */
-export const CourseProgressEdit = ( props ) => {
+export const ProgressBarEdit = ( props ) => {
 	const {
 		className,
 		barColor,
@@ -124,4 +124,4 @@ export default compose(
 			probeKey: 'primaryColor',
 		},
 	} )
-)( CourseProgressEdit );
+)( ProgressBarEdit );
