+++ assets/blocks/progress-bar-block/progress-bar.scss
@@ -1,0 +1,1 @@
+@import '../../shared/blocks/progress-bar/progress-bar';
