+++ assets/blocks/single-course-style.scss
@@ -1,2 +1,2 @@
 @import 'course-outline/course-outline';
-@import 'course-progress-block/course-progress';
+@import 'progress-bar-block/progress-bar';
