+++ assets/blocks/single-course.js
@@ -3,7 +3,7 @@
  */
 import registerSenseiBlocks from './register-sensei-blocks';
 import TakeCourseBlock from './take-course-block';
-import CourseProgressBlock from './course-progress-block';
+import ProgressBarBlock from './progress-bar-block';
 import { OutlineBlock, LessonBlock, ModuleBlock } from './course-outline';
 import ConditionalContentBlock from './conditional-content-block';
 import ViewResults from './view-results-block';
@@ -13,7 +13,7 @@ registerSenseiBlocks( [
 	ModuleBlock,
 	LessonBlock,
 	TakeCourseBlock,
-	CourseProgressBlock,
+	ProgressBarBlock,
 	ConditionalContentBlock,
 	ViewResults,
 ] );
